fix: updated Mapper, Enum, Entity, DTOs, Service and Repository
