+++ src/main/java/com/antonio/taskmanager/controller/TaskController.java
@@ -1,0 +1,10 @@
+package com.antonio.taskmanager.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/tasks")
+public class TaskController {
+    
+}
