+++ src/main/java/com/antonio/taskmanager/enums/Priority.java
@@ -1,7 +1,17 @@
 package com.antonio.taskmanager.enums;
 
 public enum Priority {
-    LOW,
-    MEDIUM,
-    HIGH
+    LOW("Baixa"),
+    MEDIUM("Média"),
+    HIGH("Alta");
+
+    private final String description;
+
+    Priority(String description) {
+        this.description = description;
+    }
+
+    public String getDescription() {
+        return description;
+    }
 }
